Remove global variable.
By the way, is there a reason you use return value instead of exceptions? It
would make the debugging of complicated expressions much easier by throwing a
description of the faulty expression. If you prefer to not use exceptions, it's
also possible to have a return type that contains a string.

I should add that currently, it look like that if you have more that one
expression, and if at least one succeded you don't get any errors.